requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
opentelemetry-api==1.17.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-propagator-b3==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
prometheus-client==0.16.0
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
